refactor: Extract types from Routes
